Create a "show" LiveView for URLMappings
As a part of developing the overall flow of going from submitting a
destination URL to providing a shortened URL, this change:

- Renames the previous `ShortenerLive` to `ShortenerLive.New`
- Create a new `ShortenerLive.Show` LiveView which will be responsible
  for displaying the shortened URL

This separation isn't strictly necessary and may be something I
revisit-it's possibly we could rely on `live_action` and/or other state
to determine which view to display. Some of the motiviation for
separating is:

- separation of concerns
- more easily relying on URLs to determine state

This work has helped set the flow and interaction between LiveViews
while avoiding creating the schema and storing items in the database,
pushing (almost?) beyond the point where that makes sense.