Define UrlMapping schema
Create a database-backed schema for representing the mapping of a
destination URL to a shortened slug that we can use to redirect users to
their target destination.

This commit the UrlMapping schema, a Shortener context for managing
mappings, and updates `ShortenerLive.New` to leverage the new context
functions to create UrlMappings.

This does not include using UrlMappings to handle the redirection yet.