Add support for tracking shortened URL usage
Create an `Event` schema to track usage of our shortened URLs. For now,
events are simply created when a user is redirects for a given `slug`.

This choice to create an `Event` schema is to enable tracking additional
data should there be a need (e.g., browser type, user location, etc.).