Update event creation logic in RedirectController
Update the event record creation step in RedirectController to be done
in a separate process via `Task.Supervisor.start_child`. This should
enable faster redirects for users and improve resiliency as the event
creation step failure will be isolated to its own process.

`Task.Supervisor.start_child` enables a fire-and-forget approach,
starting a Task under a supervisor, but not waiting for the job to
complete or tracking the exist status. This felt like the right current
trade off as event tracking/analytics feels more like a nice-to-have and
we would rather redirect a user quickly than wait for the event to be
created.

In the long-run, we may want to consider an alternative approach that
avoids the overhead of having to spin up a process on-demand (existing
process, pooling) as well as possibly bulk-writing events to be more
efficient.

This change required creating a helper function to wait for the
supervisor's children to complete before we check for results or exit
the tests.