Update styling of ShortenerLive.Show
Update the style of the shortener show page to leverage two disabled
input fields, taking inspiration from TinyURL.

As a part of re-working this, I found I wanted to share generating the
shortened URL more than I needed the `slug_link` component. After
finding the `Phoenix.VerifiedRoutes.url`, I realized I did not need to
make the URLs by hand, so I switched to use this. As a result of the
switch, the `slug_link` component was only used in one spot, so I
removed it altogether.