Handle non-unique slug constraint in URL mapping
The chance a slug collision is 64^6 which is 68.7 billion.

This is because there are 64 character options from
`Base.url_encode64/0`[^1] and our current slug length is 6 digits.

While the chances of collision are small and having a simple "something
went wrong, please retry" message is probably good enough, I wanted to
explore automatically retrying the URL mapping creation process if there
was a constraint issue.

The resulting code is definitely harder to parse. In order to keep the
test, we also have unnecessary arguments which are confusing. I am
leaving this here for discussion and I would want to discuss with the
team:

- do we want to auto-retry? give an error message?
- are these tests worth keeping around?
  They were useful to validate the implementation and I am a bit worried
  about regressions, but they require adding a default argument to
  `Shortener.create_url_mapping` which I don't really like.

[^1]: https://github.com/elixir-lang/elixir/blob/6730d669fb319411f8e411d4126f1f4067ef9231/lib/elixir/lib/base.ex#L100